[refactor][사다리-step4] BaseName - Naem클래스로 rename
