+++ src/main/java/ladder/domain/core/line/Line.java
@@ -1,7 +1,7 @@
 package ladder.domain.core.line;
 
 
-import ladder.domain.core.line.name.BaseName;
+import ladder.domain.core.line.name.Name;
 import ladder.domain.core.line.name.Names;
 import ladder.ui.result.DisplayResult;
 
@@ -21,7 +21,7 @@ public interface Line {
 
     int getWidth();
 
-    default Names<? extends BaseName> getNames(){
+    default Names<? extends Name> getNames(){
         throw new UnsupportedOperationException();
     }
