+++ src/main/java/ladder/domain/core/line/name/LadderResult.java
@@ -1,6 +1,6 @@
 package ladder.domain.core.line.name;
 
-final class LadderResult extends BaseName {
+final class LadderResult extends Name {
     public LadderResult(String name) {
         super(name);
         verifyMaxLength();
