+++ src/main/java/ladder/domain/core/line/name/LadderResultLine.java
@@ -8,7 +8,7 @@ import ladder.ui.result.DisplayResult;
 import ladder.ui.result.NamesDisplayResult;
 
 import static java.util.stream.Collectors.toList;
-import static ladder.domain.core.line.name.BaseName.verifyBlankName;
+import static ladder.domain.core.line.name.Name.verifyBlankName;
 import static ladder.domain.core.line.name.LadderResultLine.LadderResulNames.ladderResulNames;
 
 public class LadderResultLine implements Line {
