+++ src/main/java/ladder/domain/core/line/name/Name.java
@@ -1,6 +1,6 @@
 package ladder.domain.core.line.name;
 
-public abstract class BaseName {
+public class Name {
     static final String ERROR_MESSAGE_REQUIRED_NAME = "이름을 지정해 주세요.";
     static final String ERROR_MESSAGE_SUITABLE_NAME = "참여하는 사람의 이름은 알파벳(a-z), 숫자(0-9)포함 5글자 입니다.";
     static final String REGEX_SUITABLE_NAME = "[a-z][a-z0-9]{0,4}";
@@ -8,7 +8,7 @@ public abstract class BaseName {
     static final String ERROR_MESSAGE_MAX_LEN_NAME = String.format("지정된 이름의 최대 길이는 %d자 입니다.", MAX_LEN);
     protected final String name;
 
-    protected BaseName(String name) {
+    protected Name(String name) {
         verifyBlankName(name);
         this.name = name;
     }
