+++ src/test/java/ladder/domain/core/line/name/NamesLineTest.java
@@ -12,10 +12,10 @@ class NamesLineTest {
     @DisplayName("쉼표를 기준으로 공백이나 null일수 없음")
     @Test
     void blankNamesTest() {
-        assertThatIllegalArgumentException(() -> ParticipantsLine.of(null), BaseName.ERROR_MESSAGE_REQUIRED_NAME);
-        assertThatIllegalArgumentException(() -> ParticipantsLine.of(" "), BaseName.ERROR_MESSAGE_REQUIRED_NAME);
-        assertThatIllegalArgumentException(() -> LadderResultLine.of(null), BaseName.ERROR_MESSAGE_REQUIRED_NAME);
-        assertThatIllegalArgumentException(() -> LadderResultLine.of(" "), BaseName.ERROR_MESSAGE_REQUIRED_NAME);
+        assertThatIllegalArgumentException(() -> ParticipantsLine.of(null), Name.ERROR_MESSAGE_REQUIRED_NAME);
+        assertThatIllegalArgumentException(() -> ParticipantsLine.of(" "), Name.ERROR_MESSAGE_REQUIRED_NAME);
+        assertThatIllegalArgumentException(() -> LadderResultLine.of(null), Name.ERROR_MESSAGE_REQUIRED_NAME);
+        assertThatIllegalArgumentException(() -> LadderResultLine.of(" "), Name.ERROR_MESSAGE_REQUIRED_NAME);
 
     }
 
@@ -28,9 +28,9 @@ class NamesLineTest {
     @DisplayName("잘못된 이름이 있는 경우 예외 처리")
     @Test
     void suitableNameCheckTest() {
-        assertThatIllegalArgumentException(() -> ParticipantsLine.of("123"), BaseName.ERROR_MESSAGE_SUITABLE_NAME);
-        assertThatIllegalArgumentException(() -> ParticipantsLine.of("a12345"), BaseName.ERROR_MESSAGE_SUITABLE_NAME);
-        assertThatIllegalArgumentException(() -> LadderResultLine.of("a12345"), BaseName.ERROR_MESSAGE_MAX_LEN_NAME);
+        assertThatIllegalArgumentException(() -> ParticipantsLine.of("123"), Name.ERROR_MESSAGE_SUITABLE_NAME);
+        assertThatIllegalArgumentException(() -> ParticipantsLine.of("a12345"), Name.ERROR_MESSAGE_SUITABLE_NAME);
+        assertThatIllegalArgumentException(() -> LadderResultLine.of("a12345"), Name.ERROR_MESSAGE_MAX_LEN_NAME);
     }
 
     @DisplayName("참가자들는 쉼표를 기준으로 구분함")
